Refactored to extract utils from locationator.py
